JSON Serialization and Deserialization - Part 1
- Moved all JSON related methods into the JSONHelper
- Simplified some JSON
- When serializing variables, the JSON Helper will use the variables name instead of it's index to, allowing for variables to be rearranged removed/added to the node in the future and still be serialized properly. I would propose seperately that we enforce unique variable names, and potentially store them in a hash map.
- This is Part 1 of the commit, which uses the current data structure, part 2, would suggest a few new versions of the classes to allow arbitrary serialization.
An example of this is partially working for Node and NodeGraph as they both already have JsonAdapters, which allow them to be parsed to json regardless of where they are in the data structure. Whereas NodeConnection and NodeVariable can only be parsed when a Node / Graph is present which means they can't be serialized anywhere.

- Note I made bumpUpIndexableID public, obviously not a good idea for now, but I'll post my suggestion for this in another PR.